edit README.md

+++ src/controllers/courses.js
@@ -18,9 +18,13 @@ router.get('/', (req, res) => {
 
 // GET specific course
 router.get('/:id', (req, res) => {
-    Course.findOne({_id: req.params.id}).then(result => {
-        res.json(result);
-    })
+    if (req.user == null){
+        res.json({message: 'must be logged in'})
+    } else {
+        Course.findOne({_id: req.params.id}).then(result => {
+            res.json(result);
+        })
+    }
 })  // WORKS
 
 // POST new course
@@ -39,21 +43,27 @@ router.post('/', (req,res) => {
 
 // PUT endpoint
 router.put('/:id', (req,res) => {
-
-    Course.findByIdAndUpdate(req.params.id, req.body)
-        .then(result => {
-            res.json(result)
-            console.log('Object Updated Successfully')
-    })
+    if (req.user == null){
+        res.json({message: 'must be logged in'})
+    } else {
+        Course.findByIdAndUpdate(req.params.id, req.body)
+            .then(result => {
+                res.json(result)
+                console.log('Object Updated Successfully')
+        })
+    }
 }) // get again to see updated
 
 // DELETE endpoint
 router.delete('/:id', (req,res) => {
-
-    Course.findByIdAndRemove(req.params.id).then(result => {
-        res.json(result)
-        console.log('Object Deleted Successfully')
-    })
+    if (req.user == null){
+        res.json({message: 'must be logged in'})
+    } else {
+        Course.findByIdAndRemove(req.params.id).then(result => {
+            res.json(result)
+            console.log('Object Deleted Successfully')
+        })
+    }
 }) // get again to see updated
 
 module.exports = router;
